+++ app/commands/characters_context/pathfinder2/spells/add_command.rb
@@ -11,10 +11,18 @@ module CharactersContext
             required(:character).filled(type?: ::Pathfinder2::Character)
             required(:feat).filled(type?: ::Pathfinder2::Feat)
             optional(:level).filled(:integer)
+            optional(:innate).filled(:bool)
+            optional(:focus).filled(:bool)
           end
         end
 
         private
+
+        def do_prepare(input)
+          input[:value] = {}
+          input[:value] = { 'innate' => input[:innate] } if input.key?(:innate)
+          input[:value] = { 'focus' => input[:focus] } if input.key?(:focus)
+        end
 
         def do_persist(input)
           spontaneous_caster = ::Pathfinder2::Character::SPONTANEOUS_CASTERS.include?(input[:character].data.main_class)
@@ -23,7 +31,7 @@ module CharactersContext
             character: input[:character],
             feat: input[:feat],
             ready_to_use: spontaneous_caster,
-            value: spontaneous_caster ? { input[:level].to_s => { 'selected_count' => 1 } } : {}
+            value: spontaneous_caster ? input[:value].merge({ input[:level].to_s => { 'selected_count' => 1 } }) : input[:value]
           )
 
           { result: result }